requirements.txt
pandas
scikit-learn
mlflow
dvc
joblib
pyyaml
requests
numpy